+++ src/main/java/com/infoShare/calog/domain/MinorCategory/MinorCategoryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MinorCategoryRepository extends JpaRepository<MinorCategory,Long> {
+public interface MinorCategoryRepository extends JpaRepository<MinorCategory,Short> {
 }
